small modifs from windows compiler warnings

+++ include/vinecopulib/bicop/implementation/kernel.ipp
@@ -58,7 +58,8 @@ inline Eigen::VectorXd KernelBicop::hinv2(
 inline double
 KernelBicop::parameters_to_tau(const Eigen::MatrixXd &parameters) {
     set_parameters(parameters);
-    Eigen::Matrix<double, Eigen::Dynamic, 2> U = tools_stats::ghalton(1e3, 2);
+    Eigen::Matrix<double, Eigen::Dynamic, 2> U =
+        tools_stats::ghalton((size_t) 1e3, (size_t) 2);
     U.col(1) = hinv1(U);
     return tools_stats::pairwise_tau(U);
 }
